Update `skill_backend_launcher`
Added possibility to set True/False the launch of `ngrok`, `node-red`
and `azure`

+++ script/Skill/skill_backend_launcher.py
@@ -2,6 +2,14 @@
 import subprocess, os, threading, signal
 import rospy, rospkg
 
+# Open ROS Skill Server in a Separate Thread
+threading.Thread(target=lambda: rospy.init_node('skill_launcher', disable_signals=True)).start()
+
+# Get Launch Parameters
+launch_azure    = rospy.get_param('/skill_backend_launcher/launch_azure', False)
+launch_ngrok    = rospy.get_param('/skill_backend_launcher/launch_ngrok', False)
+launch_node_red = rospy.get_param('/skill_backend_launcher/launch_node_red', False)
+
 # Get Path to Package
 rospack = rospkg.RosPack()
 path = rospack.get_path('alexa_conversation')
@@ -9,15 +17,10 @@ path = rospack.get_path('alexa_conversation')
 # Set Path to ngrok
 NGROK_PATH = '/home/davide/Documenti/Programmi/ngrok/'
 
-# Open ROS Skill Server in a Separate Thread
-threading.Thread(target=lambda: rospy.init_node('skill_launcher', disable_signals=True)).start()
-
-# Launch Node-RED and ngrok in a Separate Thread Terminal
-NODE_RED = subprocess.Popen('gnome-terminal -- "node-red"', shell=True)
-NGROK    = subprocess.Popen('gnome-terminal -e "./ngrok http 7071"', cwd=f'{NGROK_PATH}', shell=True)
-
-# Launch Azure Functions
-# AZURE = subprocess.Popen('func start', cwd=f'{path}/AzureFunctions/', shell=True)
+# Launch Node-RED, ngrok and Azure in Separate Terminals
+if launch_ngrok:    NGROK    = subprocess.Popen('gnome-terminal -e "./ngrok http 7071"', cwd=f'{NGROK_PATH}', shell=True)
+if launch_node_red: NODE_RED = subprocess.Popen('gnome-terminal -- "node-red"', shell=True)
+if launch_azure:    AZURE    = subprocess.Popen('func start', cwd=f'{path}/AzureFunctions/', shell=True)
 
 def handle_signal(sig, frame):
 
@@ -25,11 +28,11 @@ def handle_signal(sig, frame):
     print("\nProgram Interrupted. Killing Opened Terminal...\n")
 
     # Kill Node-RED and ngrok
-    os.system('killall node-red')
-    os.system('killall ngrok')
+    if launch_ngrok:    os.system('killall node-red')
+    if launch_node_red: os.system('killall ngrok')
 
-    # Kill Azure Functions
-    # AZURE.wait()
+    # Wait for Azure Functions
+    if launch_azure:    AZURE.wait()
 
     print("\nDone\n")
     exit(0)
